Web_Application/requirements.txt
flask
torch
torchvision
Pillow